+++ examples/notifications-style/src/main/java/com/deltadna/android/sdk/notifications/example/ExampleActivity.java
@@ -18,11 +18,13 @@ package com.deltadna.android.sdk.notifications.example;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
 
 import com.deltadna.android.sdk.DDNA;
 import com.deltadna.android.sdk.notifications.DDNANotifications;
+import com.google.firebase.iid.FirebaseInstanceId;
 
 public class ExampleActivity extends AppCompatActivity {
     
@@ -34,9 +36,13 @@ public class ExampleActivity extends AppCompatActivity {
         
         DDNA.instance().startSdk();
         
+        final String id = DDNA.instance().getUserId();
+        Log.d(BuildConfig.LOG_TAG, "User id: " + id);
         ((TextView) findViewById(R.id.user_id)).setText(getString(
                 R.string.user_id,
-                DDNA.instance().getUserId()));
+                id));
+        
+        showRegistrationToken();
     }
     
     @Override
@@ -59,6 +65,8 @@ public class ExampleActivity extends AppCompatActivity {
      */
     public void onRegister(View view) {
         DDNANotifications.register(this);
+
+        showRegistrationToken();
     }
     
     /**
@@ -73,12 +81,12 @@ public class ExampleActivity extends AppCompatActivity {
     
     public void onNotificationOpened(View view) {
         // pretend the user opened a push notification
-        DDNA.instance().recordNotificationOpened();
+        DDNA.instance().recordNotificationOpened(false, Bundle.EMPTY);
     }
     
     public void onNotificationDismissed(View view) {
         // pretend the user dismissed a push notification
-        DDNA.instance().recordNotificationDismissed();
+        DDNA.instance().recordNotificationDismissed(Bundle.EMPTY);
     }
     
     public void onStopSdk(View view) {
@@ -87,5 +95,13 @@ public class ExampleActivity extends AppCompatActivity {
     
     public void onStartSdk(View view) {
         DDNA.instance().startSdk();
+    }
+    
+    private void showRegistrationToken() {
+        final String token = DDNA.instance().getRegistrationId();
+        Log.d(BuildConfig.LOG_TAG, "Registration token: " + token);
+        ((TextView) findViewById(R.id.registration_token)).setText(getString(
+                R.string.registration_token,
+                token));
     }
 }