+++ examples/notifications/src/main/java/com/deltadna/android/sdk/notifications/example/ExampleActivity.java
@@ -16,21 +16,16 @@
 
 package com.deltadna.android.sdk.notifications.example;
 
-import android.content.BroadcastReceiver;
 import android.os.Bundle;
-import android.support.v4.content.LocalBroadcastManager;
 import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 import android.view.View;
-
 import android.widget.TextView;
 
 import com.deltadna.android.sdk.DDNA;
 import com.deltadna.android.sdk.notifications.DDNANotifications;
 
 public class ExampleActivity extends AppCompatActivity {
-    
-    private LocalBroadcastManager broadcasts;
-    private BroadcastReceiver receiver;
     
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -40,32 +35,17 @@ public class ExampleActivity extends AppCompatActivity {
         
         DDNA.instance().startSdk();
         
+        final String id = DDNA.instance().getUserId();
+        Log.d(BuildConfig.LOG_TAG, "User id: " + id);
         ((TextView) findViewById(R.id.user_id)).setText(getString(
                 R.string.user_id,
-                DDNA.instance().getUserId()));
+                id));
         
-        broadcasts = LocalBroadcastManager.getInstance(this);
-        
-        /*
-         * Create the receiver and register it so that we can get notifications
-         * when the token changes.
-         */
-        receiver = new ExampleReceiver(
-                (TextView) findViewById(R.id.registration_token));
-        broadcasts.registerReceiver(
-                receiver,
-                DDNANotifications.FILTER_TOKEN_RETRIEVAL);
+        showRegistrationToken();
     }
     
     @Override
     public void onDestroy() {
-        /**
-         * This activity will no longer be doing anything useful with the
-         * token, however you may want to have a receiver that stays registered
-         * for the entire lifecycle of the game.
-         */
-        broadcasts.unregisterReceiver(receiver);
-        
         DDNA.instance().stopSdk();
         
         super.onDestroy();
@@ -84,6 +64,8 @@ public class ExampleActivity extends AppCompatActivity {
      */
     public void onRegister(View view) {
         DDNANotifications.register(this);
+        
+        showRegistrationToken();
     }
     
     /**
@@ -98,12 +80,12 @@ public class ExampleActivity extends AppCompatActivity {
     
     public void onNotificationOpened(View view) {
         // pretend the user opened a push notification
-        DDNA.instance().recordNotificationOpened();
+        DDNA.instance().recordNotificationOpened(false, Bundle.EMPTY);
     }
     
     public void onNotificationDismissed(View view) {
         // pretend the user dismissed a push notification
-        DDNA.instance().recordNotificationDismissed();
+        DDNA.instance().recordNotificationDismissed(Bundle.EMPTY);
     }
     
     public void onStopSdk(View view) {
@@ -112,5 +94,13 @@ public class ExampleActivity extends AppCompatActivity {
     
     public void onStartSdk(View view) {
         DDNA.instance().startSdk();
+    }
+    
+    private void showRegistrationToken() {
+        final String token = DDNA.instance().getRegistrationId();
+        Log.d(BuildConfig.LOG_TAG, "Registration token: " + token);
+        ((TextView) findViewById(R.id.registration_token)).setText(getString(
+                R.string.registration_token,
+                token));
     }
 }