+++ library-notifications/src/test/java/com/deltadna/android/sdk/notifications/NotificationInfoTest.kt
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2017 deltaDNA Ltd. All rights reserved.
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.deltadna.android.sdk.notifications
+
+import com.google.common.truth.Truth.assertThat
+import com.nhaarman.mockito_kotlin.mock
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.RuntimeEnvironment
+import org.robolectric.annotation.Config
+
+@RunWith(RobolectricTestRunner::class)
+@Config(manifest = "src/main/AndroidManifest.xml")
+class NotificationInfoTest {
+    
+    @Test
+    fun construction() {
+        val id = 1
+        val message = PushMessage(
+                RuntimeEnvironment.application,
+                "from",
+                emptyMap())
+        
+        NotificationInfo(id, message).apply {
+            assertThat(this.id).isEqualTo(id)
+            assertThat(this.message).isEqualTo(message)
+        }
+    }
+}